+++ CHANGES.md
@@ -1,3 +1,6 @@
+# unreleased
+- Support nested `let..in` for `[%sedlex.regexp?]` definitions
+
 # 3.7 (2025-10-06)
 - Update to unicode 17.0.0